+++ game_shared/util_shared.cpp
@@ -205,7 +205,7 @@ void UTIL_StringToVector( float *pVector, const char *pString, int* componentsRe
 
 	for( j = 0; j < 3; j++ )			// lifted from pr_edict.c
 	{
-		pVector[j] = atof( pfront );
+		pVector[j] = (float)atof( pfront );
 		componentsParsed++;
 
 		while( *pstr && *pstr != ' ' )
